Allow OPENCASCADE_BUILD_DIR to re-use built files
This allows more than one Alien package to use the same build
dir, so that the Foundation doesn't need re-built for every
piece of opencascade, causing an exponential build time
increase when the entire collection of libs is installed.

This does require the user's participation, but they'll see
it in the message shown telling them they need to manually
download the source file.